requirements.txt
absl-py==2.1.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
exceptiongroup==1.2.2
fastapi==0.103.2
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.66.1
h5py==3.11.0
idna==3.8
importlib_metadata==8.4.0
Jinja2==3.1.4
keras==2.13.1
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
minio==7.2.8
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.18
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.1
